requirements.txt
alive_progress==3.1.1
fire==0.1.3
ignite==1.1.0
imutils==0.5.4
matplotlib==3.6.3
networkx==2.8.8
numpy==1.23.1
pandas==1.5.2
Pillow==9.5.0
pytorch_ignite==0.4.10
scikit_learn==0.23.1
scipy==1.8.0
skimage==0.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.0
